feat: update some code

+++ src/controllers/auth.ts
@@ -19,7 +19,7 @@ export async function signup(req: Request, res: Response) {
 
   ctx.services.emailService.sendEmailLink(user.email, token.tokenCode, user.firstName, "activation");
 
-  res.status(201).send({
+  return res.status(201).send({
     status: "success",
     data: {
       id: user.id,
@@ -58,7 +58,7 @@ export async function activateAccount(req: Request, res: Response) {
 
   checkTokenExpiration(token);
 
-  const user = await ctx.db.userRepository.findUserByToken(token);
+  const user = await ctx.db.userRepository.findById(token.userId);
 
   user.active = true;
   await ctx.db.userRepository.save(user);
@@ -100,7 +100,7 @@ export async function sendRecoverPasswordLink(req: Request, res: Response) {
 
   const user = await ctx.db.userRepository.findByEmail(validBody.email);
 
-  if (user) {
+  if (user && user.active) {
     await ctx.db.tokenRepository.removeExistingTokenIfExists(user.id);
 
     const token = await ctx.db.tokenRepository.saveToken(user.id);
@@ -108,7 +108,7 @@ export async function sendRecoverPasswordLink(req: Request, res: Response) {
     ctx.services.emailService.sendEmailLink(user.email, token.tokenCode, user.firstName, "recovery");
   }
 
-  res.status(200).send({
+  return res.status(200).send({
     status: "success",
     data: null,
   });
@@ -121,7 +121,7 @@ export async function verifyRecoverPasswordLink(req: Request, res: Response) {
   const token = await ctx.db.tokenRepository.findTokenByCode(tokenCode);
   checkTokenExpiration(token);
 
-  await ctx.db.userRepository.findUserByToken(token);
+  await ctx.db.userRepository.findById(token.userId);
 
   return res.status(200).send({
     status: "success",
@@ -136,7 +136,7 @@ export async function recoverPassword(req: Request, res: Response) {
   const token = await ctx.db.tokenRepository.findTokenByCode(tokenCode);
 
   checkTokenExpiration(token);
-  const user = await ctx.db.userRepository.findUserByToken(token);
+  const user = await ctx.db.userRepository.findById(token.userId);
 
   const requiredFields = ["password", "confirmPassword"];
   const validBody = ctx.services.validateService.requestBody<RecoverPasswordBody>(req.body, requiredFields);
@@ -145,6 +145,19 @@ export async function recoverPassword(req: Request, res: Response) {
   user.password = await bcrypt.hash(validBody.password, 12);
   await ctx.db.userRepository.save(user);
   await ctx.db.tokenRepository.remove(token);
+
+  return res.status(200).send({
+    status: "success",
+    data: null,
+  });
+}
+
+export async function logout(req: Request, res: Response) {
+  const { ctx } = req;
+
+  const session = await ctx.db.sessionRepository.getValidSession(ctx.signature!.sessionId);
+
+  session!.active = false;
 
   return res.status(200).send({
     status: "success",
